+++ src/core/coms/server.ts
@@ -35,7 +35,6 @@
 import { EventEmitter } from "node:events";
 import * as fs from "node:fs";
 import * as net from "node:net";
-import * as os from "node:os";
 import * as path from "node:path";
 import { createHash, randomBytes } from "node:crypto";
 import { PeerRegistry } from "./registry.ts";